pyproject.toml
[tool.poetry]
name = "gflashcards"
version = "0.1.0"
description = "Google Sheets-based flashcard maker"
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3"
jupyter = "^1.0"
notebook = "^5.6"
google-api-python-client = "^1.7"
oauth2client = "^4.1"
markdown = "^2.6"
namedlist = "^1.7"
pillow = "^5.2"
requests = "^2.19"

[tool.poetry.dev-dependencies]
